string 테스트 추가. 빈 문자열을 사용할 때 발생하는 오류 수정(in Buffer).

+++ msgpack_wrapper/Buffer.h
@@ -13,7 +13,7 @@ public:
 	Buffer(int reserve = 1024);
 	virtual ~Buffer();
 
-	void write(const char* buf, unsigned int len);
+	void write(const char* buf, size_t len);
 
 	Writer Put();
 	Writer PutArray(size_t size);
